pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nltosql"
version = "0.1.0"
description = "Natural Language to SQL with agentic self-correction via LangGraph."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "streamlit>=1.45.0",
    "langgraph>=1.0.0",
    "langchain-ollama>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.9.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
